+++ src/function/relational/equal.js
@@ -44,12 +44,15 @@ export const createEqual = /* #__PURE__ */ factory(name, dependencies, ({ typed,
    *
    *    const a = math.unit('50 cm')
    *    const b = math.unit('5 m')
+   *    math.equal(a, b)             // returns false
+   *
+   *    const a = math.unit('500 cm')
+   *    const b = math.unit('5 m')
    *    math.equal(a, b)             // returns true
    *
    *    const c = [2, 5, 1]
    *    const d = [2, 7, 1]
-   *
-   *    math.equal(c, d)             // returns [true, false, true]
+   *    math.equal(c, d)             // returns [true, false, true]...
    *    math.deepEqual(c, d)         // returns false
    *
    *    math.equal("1000", "1e3")    // returns true
@@ -58,6 +61,16 @@ export const createEqual = /* #__PURE__ */ factory(name, dependencies, ({ typed,
    * See also:
    *
    *    unequal, smaller, smallerEq, larger, largerEq, compare, deepEqual, equalText
+   *
+   * History:
+   *
+   *    v13    Handle bigints
+   *    v11.6  Support matrix broadcasting
+   *    v4     Compare strings by their numerical value
+   *    v0.24  Handle `null` and `undefined`
+   *    v0.23  Compare collections elementwise
+   *    v0.20  Compare floating-point numbers within epsilon, allowing roundoff
+   *    v0.2   Created
    *
    * @param  {number | BigNumber | bigint | boolean | Complex | Unit | string | Array | Matrix} x First value to compare
    * @param  {number | BigNumber | bigint | boolean | Complex | Unit | string | Array | Matrix} y Second value to compare