+++ src/type/fraction/function/fraction.js
@@ -37,6 +37,14 @@ export const createFraction = /* #__PURE__ */ factory(name, dependencies, ({ typ
    *
    *    bignumber, number, string, unit
    *
+   * History:
+   *
+   *    v2     Created
+   *    v3     Added conversion from BigNumber
+   *    v11.8  Added conversion from Unit
+   *    v13    Added conversion from bigint
+   *    v14    Move to bigint-based fraction.js@5; construct from two bigints
+   *
    * @param {number | string | Fraction | BigNumber | bigint | Unit | Array | Matrix} [args]
    *            Arguments specifying the value, or numerator and denominator of
    *            the fraction